Change skill power int to int64

+++ Lib9c/Battle/CriticalHelper.cs
@@ -12,22 +12,22 @@ namespace Nekoyume.Battle
     {
         private const decimal MinimumDamageMultiplier = 1m;
 
-        public static int GetCriticalDamage(CharacterBase caster, int originalDamage)
+        public static long GetCriticalDamage(CharacterBase caster, long originalDamage)
         {
             var critMultiplier =
                 Math.Max(
                     MinimumDamageMultiplier,
                     CharacterBase.CriticalMultiplier + (caster.CDMG / 10000m));
-            return (int)(originalDamage * critMultiplier);
+            return (long)(originalDamage * critMultiplier);
         }
 
-        public static int GetCriticalDamageForArena(ArenaCharacter caster, int originalDamage)
+        public static long GetCriticalDamageForArena(ArenaCharacter caster, long originalDamage)
         {
             var critMultiplier =
                 Math.Max(
                     MinimumDamageMultiplier,
                     ArenaCharacter.CriticalMultiplier + (caster.CDMG / 10000m));
-            return (int)(originalDamage * critMultiplier);
+            return (long)(originalDamage * critMultiplier);
         }
     }
 }
