+++ Lib9c/Model/Skill/SkillCustomField.cs
@@ -4,6 +4,6 @@ namespace Nekoyume.Model.Skill
     public struct SkillCustomField
     {
         public int BuffDuration { get; set; }
-        public int BuffValue { get; set; }
+        public long BuffValue { get; set; }
     }
 }
